Standard methods for estimated models

coef and vcov for hlme, lcmm, mutlcmm, Jointlcmm, mpjlcmm, externSurv, and externX models, fixef, ranef, fitted and residuals methods for estimated hlme, lcmm, mutlcmm and Jointlcmm models.

Usage

# S3 method for class 'hlme'
coef(object, ...)

Arguments

object

an object of class hlme, lcmm, multlcmm or Jointlcmm

...

other arguments. There are ignored in these functions.

Value

For coef, the vector of the estimates.

For vcov, the variance-covariance matrix of the estimates.

For fixef : - for hlme, lcmm and multlcmm objects, a list containing the fixed effects estimates in the class-membership model and in the longitudinal model. - for Jointlcmm objects, a list containing the fixed effects estimates in the class-membership model, the survival model and in the longitudinal model.

For ranef, a matrix (nrow=number of subjects, ncol=number of covariates with random effect) containing the individual random effects.

For fitted, a vector containing the subject-specific predictions extracted from object.

For residuals, a vector containing the subject-specific residuals extracted from object.

Author

Cecile Proust-Lima, Viviane Philipps