Variance-covariance of the estimates

This function provides the variance-covariance matrix of the estimates. vcov is an alias for it.

Usage

VarCov(x)

Arguments

x

an object of class hlme, lcmm, multlcmm, Jointlcmm or mpjlcmm

Value

a matrix containing the variance-covariance of the estimates. For the parameters of the matrix of variance-covariance of the random effects, the Cholesky transformed parameters are considered so that VarCov provides the covariance matrix of function estimates with cholesky=TRUE.

See also

Author

Cecile Proust-Lima, Viviane Philipps