Marginal predictions of the longitudinal outcome(s) in their natural scale from lcmm, Jointlcmm or multlcmm objects

The function computes the marginal predictions of the longitudinal outcome(s) in their natural scale on the individual data used for the estimation from lcmm, Jointlcmm or multlcmm objects.

Usage

fitY(x)

Arguments

x

an object inheriting from classes lcmm or multlcmm.

Value

For lcmm and Jointlcmm objects, returns a matrix with ng+1 columns containing the subject identifier and the ng class-specific marginal predicted values.

For multlcmm objects, returns a matrix with ng+2 columns containing the subject identifier, the outcome indicator and the ng class-specific predicted values.

See also

Author

Cecile Proust-Lima, Viviane Philipps