Longitudinal data on cognitive and physical aging in the elderly

The dataset consists in a subsample of the Paquid prospective cohort study. Repeated measures cognitive measures (MMSE, IST, BVRT psychometric tests), physical dependency (HIER) and depression sympatomatology (CESD) were collected over a maximum period of 20 years along with dementia information (age at dementia diagnosis, dementia diagnosis information). Time-independent socio-demographic information is also provided (CEP, male, age_init).

Format

A data frame with 2250 observations over 500 subjects and 12 variables:

ID

subject identification number

MMSE

score at the Mini-Mental State Examination (MMSE), a psychometric test of global cognitive functioning (integer in range 0-30)

BVRT

score at the Benton Visual Retention Test (BVRT), a psychometric test of spatial memory (integer in range 0-15)

IST

score at the Isaacs Set Test (IST) truncated at 15 seconds, a test of verbal memory (integer in range 0-40)

HIER

score of physical dependency (0=no dependency, 1=mild dependency, 2=moderate dependency, 3=severe dependency)

CESD

score of a short self-report scale CES-D designed to measure depressive symptomatology in the general population (integer in range 0-52)

age

age at the follow-up visit

dem

indicator of positive diagnosis of dementia

agedem

age at dementia diagnosis for dem=1 and at last contact for dem=0

age_init

age at entry in the cohort

CEP

binary indicator of educational level (CEP=1 for subjects who graduated from primary school; CEP=0 otherwise)

male

binary indicator for gender (male=1 for men; male=0 for women)

References

Letenneur, L., Commenges, D., Dartigues, J. F., & Barberger-Gateau, P. (1994). Incidence of dementia and Alzheimer's disease in elderly community residents of southwestern France. International Journal of Epidemiology, 23 (6), 1256-61.

Examples

summary(paquid)
#>        ID             MMSE            BVRT            IST       
#>  Min.   :  1.0   Min.   : 0.00   Min.   : 0.00   Min.   : 5.00  
#>  1st Qu.:132.2   1st Qu.:25.00   1st Qu.: 9.00   1st Qu.:22.00  
#>  Median :263.0   Median :27.00   Median :11.00   Median :27.00  
#>  Mean   :256.2   Mean   :25.99   Mean   :10.78   Mean   :26.52  
#>  3rd Qu.:376.0   3rd Qu.:29.00   3rd Qu.:13.00   3rd Qu.:31.00  
#>  Max.   :500.0   Max.   :30.00   Max.   :15.00   Max.   :40.00  
#>                  NA's   :36      NA's   :300     NA's   :198    
#>       HIER            CESD             age             agedem     
#>  Min.   :0.000   Min.   : 0.000   Min.   : 66.28   Min.   :66.87  
#>  1st Qu.:1.000   1st Qu.: 2.000   1st Qu.: 75.09   1st Qu.:82.03  
#>  Median :1.000   Median : 6.000   Median : 80.57   Median :86.61  
#>  Mean   :1.272   Mean   : 8.488   Mean   : 80.53   Mean   :85.89  
#>  3rd Qu.:2.000   3rd Qu.:12.000   3rd Qu.: 85.77   3rd Qu.:89.94  
#>  Max.   :3.000   Max.   :52.000   Max.   :102.16   Max.   :99.49  
#>  NA's   :46      NA's   :146                                      
#>       dem            age_init          CEP              male       
#>  Min.   :0.0000   Min.   :65.25   Min.   :0.0000   Min.   :0.0000  
#>  1st Qu.:0.0000   1st Qu.:68.00   1st Qu.:0.0000   1st Qu.:0.0000  
#>  Median :0.0000   Median :71.92   Median :1.0000   Median :0.0000  
#>  Mean   :0.3329   Mean   :73.01   Mean   :0.7373   Mean   :0.3853  
#>  3rd Qu.:1.0000   3rd Qu.:76.96   3rd Qu.:1.0000   3rd Qu.:1.0000  
#>  Max.   :1.0000   Max.   :92.33   Max.   :1.0000   Max.   :1.0000  
#>