Plot of information functions

This function plots the information functions stemmed from a lcmm or multlcmm object with ordinal outcomes modeled via threshold links.

Usage

# S3 method for class 'ItemInfo'
plot(
  x,
  which = "ItemInfo",
  outcome = "all",
  legend.loc = "topright",
  legend = NULL,
  add = FALSE,
  shades = TRUE,
  ...
)

Arguments

x

an object inheriting from classes ItemInfo

which

character specifying the values to plot. Should be one of 'ItemInfo' for the Fisher information function of the ordinal outcomes, 'LevelInfo' for the information of each item's level or 'LevelProb' for the probability of the item's levels. Default to 'ItemInfo'.

outcome

character specifying the outcome to consider. Default to "all".

legend.loc

keyword for the position of the legend from the list "bottomright", "bottom", "bottomleft", "left", "topleft","top", "topright", "right" and "center".

legend

character or expression to appear in the legend. If no legend should be added, "legend" should be NULL.

add

logical indicating if the curves should be added to an existing plot. Default to FALSE.

shades

logical indicating if confidence intervals should be represented with shades. Default to FALSE, the confidence intervals are represented with dotted lines.

...

other parameters to be passed through to plotting functions or to legend

Author

Viviane Philipps and Cecile Proust-Lima