Plot of predicted cumulative incidences according to a profile of covariates

This function displays the predicted cause-specific cumulative incidences derived from a joint latent class model according to a profile of covariates. does. ~~

Usage

# S3 method for class 'cuminc'
plot(
  x,
  profil = 1,
  event = 1,
  add = FALSE,
  legend,
  legend.loc = "topleft",
  ...
)

Arguments

x

an object of class cuminc

profil

an integer giving the profile number for which the cumulative incidences are to be plotted.

event

an integer giving the event indicator for which the cumulative incidence are to be plotted.

add

logical indicating if the curves should be added to an existing plot. Default to FALSE.

legend

character or expression to appear in the legend. If no legend should be added, "legend" should be NULL.

legend.loc

keyword for the position of the legend from the list "bottomright", "bottom", "bottomleft", "left", "topleft","top", "topright", "right" and "center". By default, the legend is located in the top left of the plot.

...

other parameters to be passed through to plotting functions

Value

returns NULL

See also

Author

Viviane Philipps and Cecile Proust-Lima