Summary of a hlme, lcmm, Jointlcmm, multlcmm, mpjlcmm, externSurv, externX epoce or Diffepoce objects

The function provides a summary of hlme, lcmm, multlcmm and Jointlcmm estimations, or epoce and Diffepoce computations.

Usage

# S3 method for class 'lcmm'
summary(object, ...)

Arguments

object

an object inheriting from classes hlme, lcmm, multlcmm for fitted latent class mixed-effects, or class Jointlcmm, mpjlcmm for a Joint latent class mixed model or epoce or Diffepoce for predictive accuracy computations or externSurv, externX for secondary regression models.

...

further arguments to be passed to or from other methods. They are ignored in this function.

Value

For epoce or Diffepoce objects, returns NULL. For hlme, lcmm, Jointlcmm or multlcmm returns also a matrix containing the fixed effect estimates in the longitudinal model, their standard errors, Wald statistics and p-values

See also

Author

Cecile Proust-Lima, Viviane Philipps, Amadou Diakite and Benoit Liquet