Summary of models

This function provides a table summarizing the results of different models fitted by hlme, lcmm, multlcmm, Jointlcmm, mpjlcmm or externVar.

Usage

summarytable(
  m1,
  ...,
  which = c("G", "loglik", "npm", "BIC", "%class"),
  display = TRUE
)

Arguments

m1

an object of class hlme, lcmm, multlcmm, Jointlcmm, mpjlcmm, externVar or externVar.

...

further arguments, in particular other objects of class hlme, lcmm, multlcmm, Jointlcmm or mpjlcmm.

which

character vector indicating which results should be returned. Possible values are "G", "loglik", "conv", "npm", "AIC", "BIC", "SABIC", "entropy", "ICL", "ICL1", "ICL2", "%class".

display

logical indicating whether the table should be printed (the default) or not (display=FALSE)

Value

a matrix giving for each model the values of the requested indexes. By default, the number a latent classes, the log-likelihood, the number of parameters, the BIC and the posterior probability of the latent classes.

Details

Can be reported the usual criteria used to assess the fit and the clustering of the data: - maximum log-likelihood L (the higher the better) - number of parameters P, number of classes G, convergence criterion (1 = converged) - AIC (the lower the better) computed as -2L+2P - BIC (the lower the better) computed as -2L+ P log(N) where N is the number of subjects - SABIC (the lower the better) computed as -2L+ P log((N+2)/24) - Entropy (the closer to one the better) computed as 1+sum[pi_ig*log(pi_ig)]/(N*log(G)) where pi_ig is the posterior probability that subject i belongs to class g - ICL (the lower the better) computed in two ways : ICL1 = BIC - sum[pi_ig*log(pi_ig)] or ICL2 = BIC - 2*sum(log(max(pi_ig)), where the max is taken over the classes for each subject. - %class computed as the proportion of each class based on c_ig

See also

Author

Cecile Proust-Lima, Viviane Philipps