Wrapper to the Fortran subroutines computing the log-likelihood

Log-likelihood of hlme, lcmm, multlcmm, Jointlcmm and mpjlcmm models. The argument's specification is not straightforward, so these functions are usually not directly used.

Usage

loglikhlme(
  b,
  Y0,
  X0,
  prior0,
  pprior0,
  idprob0,
  idea0,
  idg0,
  idcor0,
  ns0,
  ng0,
  nv0,
  nobs0,
  nea0,
  nmes0,
  idiag0,
  nwg0,
  ncor0,
  npm0,
  fix0,
  nfix0,
  bfix0
)

logliklcmm(
  b,
  Y0,
  X0,
  prior0,
  pprior0,
  idprob0,
  idea0,
  idg0,
  idcor0,
  ns0,
  ng0,
  nv0,
  nobs0,
  nea0,
  nmes0,
  idiag0,
  nwg0,
  ncor0,
  npm0,
  epsY0,
  idlink0,
  nbzitr0,
  zitr0,
  minY0,
  maxY0,
  ide0,
  fix0,
  nfix0,
  bfix0
)

loglikmultlcmm(
  b,
  Y0,
  X0,
  prior0,
  pprior0,
  idprob0,
  idea0,
  idg0,
  idcor0,
  idcontr0,
  ny0,
  ns0,
  ng0,
  nv0,
  nobs0,
  nea0,
  nmes0,
  idiag0,
  nwg0,
  ncor0,
  nalea0,
  npm0,
  epsY0,
  idlink0,
  nbzitr0,
  zitr0,
  uniqueY0,
  indiceY0,
  nvalSPLORD0,
  fix0,
  nfix0,
  bfix0,
  methInteg0,
  nMC0,
  dimMC0,
  seqMC0,
  chol0
)

loglikJointlcmm(
  b,
  Y0,
  X0,
  prior0,
  pprior0,
  tentr0,
  tevt0,
  devt0,
  ind_survint0,
  idprob0,
  idea0,
  idg0,
  idcor0,
  idcom0,
  idspecif0,
  idtdv0,
  idlink0,
  epsY0,
  nbzitr0,
  zitr0,
  uniqueY0,
  nvalSPL0,
  indiceY0,
  typrisq0,
  risqcom0,
  nz0,
  zi0,
  ns0,
  ng0,
  nv0,
  nobs0,
  nmes0,
  nbevt0,
  nea0,
  nwg0,
  ncor0,
  idiag0,
  idtrunc0,
  logspecif0,
  npm0,
  fix0,
  nfix0,
  bfix0
)

loglikmpjlcmm(
  b,
  K0,
  ny0,
  nbevt0,
  ng0,
  ns0,
  Y0,
  nobs0,
  X0,
  nv0,
  Xns0,
  nv20,
  prior0,
  Tentr0,
  Tevt0,
  Devt0,
  ind_survint0,
  idnv0,
  idnv20,
  idspecif0,
  idlink0,
  epsY0,
  nbzitr0,
  zitr0,
  uniqueY0,
  nvalSPL0,
  indiceY0,
  typrisq0,
  risqcom0,
  nz0,
  zi0,
  nmes0,
  nea0,
  nw0,
  ncor0,
  nalea0,
  idiag0,
  idtrunc0,
  logspecif0,
  npm0,
  fix0,
  contrainte0,
  nfix0,
  bfix0
)

Arguments

b

the vector of estimated parameters (length npm0)

Y0

the observed values of the outcome(s) (length nobs0)

X0

the observed values of all covariates included in the model (dim nob0 * nv0)

prior0

the prior latent class (length ns0)

pprior0

the prior probabilty of each latent class (dim ns0 * ng0)

idprob0

indicator of presence in the class membership submodel (length nv0)

idea0

indicator of presence in the random part of the longitudinal submodel (length nv0)

idg0

indicator of presence in the fixed part of the longitudinal submodel (length nv0)

idcor0

indicator of presence in the correlation part of the longitudinal submodel (length nv0)

ns0

number of subjects

ng0

number of latent classes

nv0

number of covariates

nobs0

number of observations

nea0

number of random effects

nmes0

number of mesures for each subject (length ns0 or dom ns0*ny0)

idiag0

indicator of diagonal variance matrix of the random effects

nwg0

number of parameters for proportional random effects over latent classes

ncor0

number of parameters for the correlation

npm0

total number of parameters

fix0

indicator of non estimated parameter (length npm0+nfix0)

nfix0

number of non estimated parameters

bfix0

vector of non estimated parameters

epsY0

epsY values for Beta transformations

idlink0

type of transformation

nbzitr0

number of nodes for the transformations

zitr0

nodes for the transformations

minY0

minimum value for the longitudinal outcome

maxY0

maximum value for the longitudinal outcome

ide0

indicator of observed values for ordinal outcomes

idcontr0

indicator of presence as contrast in the fixed part of the longitudinal submodel (length nv0)

ny0

number of longitudinal outcomes

nalea0

number of parameters f the outcome specific random effect

uniqueY0

unique values of the longitudinal outcomes

indiceY0

correspondance between Y0 and uniqueY0

nvalSPLORD0

number of unique values for outcomes modeled with splines transformations or as ordinal outcome

methInteg0

type of integration

nMC0

number of nodes for Monte Carlo integration

dimMC0

dimension of the integration

seqMC0

sequence of integration nodes

chol0

indicator of Cholesky parameterization

tentr0

entry time for the survival submodel

tevt0

event time for the survival submodel

devt0

indicator of event for the survival submodel

ind_survint0

indicator of risk change

idcom0

indicator of presence in the survival submodel with common effect

idspecif0

indicator of presence in the survival submodel with cause-specific or class specific effect

idtdv0

indicator of 'TimeDepVar' covariate

nvalSPL0

number of unique values for outcomes modeled with splines transformations

typrisq0

type of baseline risk

risqcom0

specification of baseline risk across latent classes

nz0

number of nodes for the baseline

zi0

nodes for the baseline

nbevt0

number of events

idtrunc0

indicator of left truncation

logspecif0

indicator of logarithm parameterization

K0

number of latent processes

Xns0

the observed values of the covariates included in the survival submodel (dim ns0*nv20)

nv20

number of covariates in Xns0

Tentr0

entry time for the survival submodel (length ns0)

Tevt0

event time for the survival submodel (length ns0)

Devt0

indicator of event for the survival submodel (length ns0)

idnv0

indicator of presence in each subpart of the longitudinal models (length 4*sum(nv0))

idnv20

indicator of presence in each subpart of the survival models (length 3*nv20)

nw0

number of parameters for proportional random effects over latent classes

contrainte0

type of identifiability constraints

Value

the log-likelihood

Author

Cecile Proust-Lima, Viviane Philipps